Separated into create intent and select quote components
