+++ apps/sdk-demo/src/styles/globals.css
@@ -16,3 +16,10 @@ a {
 * {
   box-sizing: border-box;
 }
+
+pre {
+  background-color: #f4f4f4;
+  padding: 0.5rem;
+  border: 1px solid #ccc;
+  overflow-x: auto;
+}
